feat: implement letters-mongo-repository

+++ src/models/letters/use-cases/delete/delete-letter-use-case.ts
@@ -9,6 +9,13 @@ export class DeleteLetterUseCase {
 
   async execute(id: string): Promise<void> {
 
+    const letterExists = await this.letterRepository.findById(id);
+
+    if (!letterExists) {
+      throw new Error(`Letter ID: ${id} not found !`)
+    };
+
+
     await this.letterRepository.delete(id)
   }
